# Faster connected_components